# packages/web/public/openapi.yaml
openapi: 3.0.3
info:
  title: Royal Battalion API
  description: |
    REST API for the Royal Battalion gaming community management platform.
    Handles user authentication (Discord OAuth), whitelist management,
    ticket/prospect tracking, match history, server administration,
    and Discord bot integration.

    ## Versioning
    All REST endpoints are versioned and served under the `/v1` prefix, which is
    baked into the `servers[].url` base below (so the paths documented here are
    relative to `/v1`, e.g. `/users` resolves to `/v1/users`).

    A few root-level, NON-versioned endpoints are exempt (external / infra
    contracts that must not move) and declare their own root `servers` override:
    - `GET /health` -- infra/Railway health probe.
    - `GET /admins.cfg` -- fetched by the Squad game server at a fixed URL.

    ## Authentication
    Most endpoints require a JWT Bearer token obtained via the `/auth/sync` endpoint.
    The flow is:
    1. User authenticates with Discord OAuth on the frontend
    2. Frontend sends the Discord access token to `POST /auth/sync`
    3. API returns a JWT (valid 4 hours) containing userId and permissions
    4. Include the JWT in all subsequent requests as `Authorization: Bearer <token>`

    ## Permissions
    Access is controlled by a granular RBAC system with 19 permissions.
    Users with the `admin` permission bypass all permission checks.

    ## Rate Limiting
    - Global: 200 requests/minute per IP
...

...
  version: 1.0.0
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /whitelist:
    get:
      tags: [Whitelist]
      summary: List whitelist entries
      description: Returns all whitelist entries, optionally filtered by server.
      parameters:
        - name: server
          in: query
          schema:
            type: string
          description: Filter by server key (e.g., "main", "battle")
      responses:
        "200":
          description: Whitelist entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_WhitelistEntryArray"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
      x-required-permissions:
        - view:whitelist
    post:
      tags: [Whitelist]
      summary: Add whitelist entry
      description: Adds a single whitelist entry and triggers SFTP deployment.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddWhitelistEntry"
      responses:
        "201":
          description: Entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_WhitelistEntry"
        "400":
          description: Steam ID already whitelisted on this server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
      x-required-permissions:
        - manage:whitelist
  /whitelist/candidates:
    get:
      tags: [Whitelist]
      summary: Get whitelist candidates
      description: Returns users eligible for whitelist (have a whitelist-granting role but aren't yet whitelisted).
      parameters:
        - name: server
          in: query
          schema:
            type: string
            default: main
      responses:
        "200":
          description: Candidate list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/WhitelistCandidate"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
      x-required-permissions:
        - manage:whitelist
  /whitelist/bulk:
    post:
      tags: [Whitelist]
      summary: Bulk add whitelist entries
      description: Adds multiple whitelist entries (1-500) and triggers SFTP deployment. Duplicate Steam IDs are skipped.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [entries]
              properties:
                server:
                  type: string
                  default: main
                entries:
                  type: array
                  minItems: 1
                  maxItems: 500
                  items:
                    type: object
                    required: [steamId]
                    properties:
                      steamId:
                        type: string
                      name:
                        type: string
                      clan:
                        type: string
                      role:
                        type: string
                      groupId:
                        type: string
                      reason:
                        type: string
      responses:
        "201":
          description: Bulk add result (duplicates are skipped, not failed)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      created:
                        type: integer
                      skipped:
                        type: array
                        items:
                          type: object
                          properties:
                            steamId:
                              type: string
                            reason:
                              type: string
                              enum: [duplicate_existing, duplicate_in_batch]
                            existingName:
                              type: string
                              nullable: true
                            existingExpiresAt:
                              type: string
                              format: date-time
                              nullable: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
      x-required-permissions:
        - manage:whitelist
  /whitelist/bulk-update:
    post:
      tags: [Whitelist]
      summary: Bulk update whitelist entries
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids, data]
              properties:
                ids:
                  type: array
                  items:
                    type: string
                data:
                  type: object
                  properties:
                    clanId:
                      type: string
                      nullable: true
                    groupId:
                      type: string
                      nullable: true
                    expiresAt:
                      type: string
                      format: date-time
                      nullable: true
      responses:
        "200":
          description: Bulk update result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      updated:
                        type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
      x-required-permissions:
        - manage:whitelist
  /whitelist/bulk-delete:
    post:
      tags: [Whitelist]
      summary: Bulk delete whitelist entries
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids]
              properties:
                ids:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: Bulk delete result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      deleted:
                        type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
      x-required-permissions:
        - manage:whitelist
  /whitelist/{id}:
    patch:
      tags: [Whitelist]
      summary: Update whitelist entry
      description: Partially updates an existing whitelist entry and triggers SFTP deployment.
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                steamId:
                  type: string
                name:
                  type: string
                clan:
                  type: string
                clanId:
                  type: string
                  nullable: true
                role:
                  type: string
                groupId:
                  type: string
                  nullable: true
                reason:
                  type: string
                expiresAt:
                  type: string
                  nullable: true
                  format: date-time
      responses:
        "200":
          description: Updated entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_WhitelistEntry"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Entry not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      x-required-permissions:
        - manage:whitelist
    delete:
      tags: [Whitelist]
      summary: Delete whitelist entry
      description: Removes a whitelist entry and triggers SFTP deployment.
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Entry not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      x-required-permissions:
        - manage:whitelist
  /whitelist/{id}/comments:
    post:
      tags: [Whitelist]
      summary: Add whitelist comment
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                  minLength: 1
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: "#/components/schemas/WhitelistComment"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
      x-required-permissions:
        - manage:whitelist
  /whitelist/{id}/comments/{commentId}:
    delete:
      tags: [Whitelist]
      summary: Delete whitelist comment
      description: Deletes a comment scoped by both the entry id and comment id.
      parameters:
        - $ref: "#/components/parameters/ResourceId"
        - name: commentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Comment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      x-required-permissions:
        - manage:whitelist

  # ── Clans ─────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT obtained from POST /auth/sync
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NoContent:
      description: Success (no content)
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
...
    AddWhitelistEntry:
      type: object
      required: [steamId]
      properties:
        steamId:
          type: string
          minLength: 1
        server:
          type: string
          default: main
        name:
          type: string
        clan:
          type: string
        role:
          type: string
        groupId:
          type: string
        reason:
          type: string
        expiresAt:
          type: string
          format: date-time
    WhitelistCandidate:
      type: object
      properties:
        userId:
          type: string
        discordName:
          type: string
        steamId:
          type: string
        roleName:
          type: string
    ApiResponse_WhitelistEntry:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: "#/components/schemas/WhitelistEntry"
    ApiResponse_WhitelistEntryArray:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: "#/components/schemas/WhitelistEntry"

    # ── Tickets ─────────────────────────────────────────
...
    WhitelistComment:
      type: object
      properties:
        id:
          type: string
        whitelistEntryId:
          type: string
        authorId:
          type: string
        authorName:
          type: string
        text:
          type: string
        createdAt:
          type: string
          format: date-time
...